Creo JavaDoc de Producto y actualizo proyecto con todos los cambios.

+++ Calculadora-Actividad2/src/main/java/org/Calculadora/Suma.java
@@ -138,7 +138,7 @@ public class Suma {
      */
 
         public static double sumaTresReales(double num1, double num2, double num3) {
-            return num1 + num2 + num3;
+            return (double) num1 + num2 + num3;
         }
 
     /**
